Traffic Ops - test helper module now properly deletes cachegroups without relying on order of ids

+++ traffic_ops/app/lib/Test/TestHelper.pm
@@ -145,10 +145,21 @@ sub teardown {
 sub teardown_cachegroup {
 	my $self   = shift;
 	my $schema = shift;
-	my $cachegroups =
-		$schema->resultset("Cachegroup")->search( undef, { order_by => { -desc => [qw{parent_cachegroup_id secondary_parent_cachegroup_id}] } } );
-	while ( my $row = $cachegroups->next ) {
-		$row->delete();
+
+	my $cachegroups = $schema->resultset("Cachegroup");
+	while ( $cachegroups ne 0 ) {
+		while ( my $row = $cachegroups->next ) {
+			if ( $schema->resultset("Cachegroup")->count({parent_cachegroup_id => $row->id}) ne 0 ) {
+				next;
+			}
+
+			if ( $schema->resultset("Cachegroup")->count({secondary_parent_cachegroup_id => $row->id}) ne 0 ) {
+				next;
+			}
+
+			$row->delete();
+		}
+		$cachegroups = $schema->resultset("Cachegroup");
 	}
 }
